balance payment integrated

+++ lib/ui/features_user/booking/booking_controller/bookings_list_controller.dart
@@ -52,7 +52,7 @@ class BookingsListController extends GetxController {
       double balanceAmount;
       try {
         balanceAmount = bookingDetails.totalCalculatedServiceCharge! -
-            int.parse(bookingDetails.depositPayment!.amount!);
+            double.parse(bookingDetails.depositPayment!.amount!);
       } catch (e) {
         balanceAmount = bookingDetails.service!.baseCost! - 100;
       }
